re-organize views into a module
