+++ corehq/apps/data_cleaning/urls.py
@@ -1,8 +1,10 @@
 from django.urls import re_path as url
 
-from corehq.apps.data_cleaning.views import (
+from corehq.apps.data_cleaning.views.main import (
     CleanCasesMainView,
     CleanCasesSessionView,
+)
+from corehq.apps.data_cleaning.views.tables import (
     CleanCasesTableView,
 )
