+++ corehq/apps/data_cleaning/views/main.py
@@ -6,12 +6,7 @@ from django.utils.decorators import method_decorator
 from django.utils.translation import gettext_lazy, gettext as _
 
 from corehq import toggles
-from corehq.apps.data_cleaning.tables import CleanCaseTable
-from corehq.apps.domain.decorators import LoginAndDomainMixin
-from corehq.apps.domain.views import DomainViewMixin
-from corehq.apps.es import CaseSearchES
 from corehq.apps.hqwebapp.decorators import use_bootstrap5
-from corehq.apps.hqwebapp.tables.pagination import SelectablePaginatedTableView
 from corehq.apps.settings.views import BaseProjectDataView
 
 
@@ -71,19 +66,3 @@ class CleanCasesSessionView(BaseProjectDataView):
         return {
             "session_id": self.session_id,
         }
-
-
-@method_decorator([
-    use_bootstrap5,
-    toggles.DATA_CLEANING_CASES.required_decorator(),
-], name='dispatch')
-class CleanCasesTableView(LoginAndDomainMixin, DomainViewMixin, SelectablePaginatedTableView):
-    urlname = "data_cleaning_cases_table"
-    table_class = CleanCaseTable
-
-    @property
-    def session_id(self):
-        return self.kwargs['session_id']
-
-    def get_queryset(self):
-        return CaseSearchES().domain(self.domain)
